AC folder rename & AC namespacing
