+++ Assets/ActivateCode/CH.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: d5e1d91fd4460224ba25bf66262b4747
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
